X11: Discard real ConfigureNotify events
Fun fact: real ConfigureNotify events have positions relative to
the parent window, whereas synthetic ones have positions relative
to the root window.

What this commit ultimately accomplishes is preventing any Moved
events with a parent-relative position from being sent to the user,
which was something that was happening when resizing.